requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers
djangorestframework-simplejwt
drf-spectacular
psycopg2-binary
zeep
requests
aiohttp
asyncio
django-colorfield
boto3
django-storages
gunicorn
zarinpal-py-sdk
